Add support for PriceRules API.

+++ ShopifySharp/Entities/PrerequisiteValueRange.cs
@@ -5,9 +5,9 @@ namespace ShopifySharp
     public class PrerequisiteValueRange
     {
         [JsonProperty("less_than_or_equal_to")]
-        public long? LessThanOrEqualTo { get; set; }
+        public decimal? LessThanOrEqualTo { get; set; }
 
         [JsonProperty("greater_than_or_equal_to")]
-        public long? GreaterThanOrEqualTo { get; set; }
+        public decimal? GreaterThanOrEqualTo { get; set; }
     }
 }
